rename non global components non camelcase
